+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- Added: `repo.saveAndSync` command (Ctrl + Shift + S)
+
 # v0.0.1
 
 - Initial release